requirements.txt
alembic==1.8.1
apispec==5.2.2
click==8.1.3
Flask==2.2.2
Flask-JWT-Extended==4.4.4
Flask-Migrate==3.1.0
flask-smorest==0.39.0
Flask-SQLAlchemy==3.0.1
greenlet==1.1.3.post0
importlib-metadata==5.0.0
importlib-resources==5.10.0
itsdangerous==2.1.2
#Jinja2==3.1.2
Jinja2==3.0
Mako==1.2.3
MarkupSafe==2.1.1
marshmallow==3.18.0
mysql==0.0.3
mysqlclient==2.1.1
packaging==21.3
passlib==1.7.4
PyJWT==2.5.0
pyparsing==3.0.9
python-dotenv==0.21.0
SQLAlchemy==1.4.41
webargs==8.2.0
Werkzeug==2.2.2
zipp==3.9.0
